#238 additional specs, code clean up
