+++ lib/isodoc/html_function/postprocess.rb
@@ -1,34 +1,7 @@
+require "isodoc/html_function/mathvariant_to_plain"
+
 module IsoDoc::HtmlFunction
   module Html
-    MATHML = { "m" => "http://www.w3.org/1998/Math/MathML" }.freeze
-    MATHVARIANT_SPECIAL_CASE_MAPPINGS_1 = %w[bold italic sans-serif]
-      .permutation
-      .each_with_object(:sansbolditalic)
-      .map { |n, y| [n, y] }
-      .to_h
-      .freeze
-    MATHVARIANT_SPECIAL_CASE_MAPPINGS_2 = {
-      %w[bold fraktur] => :frakturbold,
-      %w[bold script] => :scriptbold,
-      %w[sans-serif bold] => :sansbold,
-      %w[sans-serif italic] => :sansitalic,
-      %w[sans-serif bold-italic] => :sansbolditalic,
-      %w[bold-sans-serif italic] => :sansbolditalic,
-      %w[sans-serif-italic bold] => :sansbolditalic,
-    }.freeze
-    MATHVARIANT_TO_PLANE_MAPPINGS = {
-      %w[double-struck] => :doublestruck,
-      %w[bold-fraktur] => :frakturbold,
-      %w[script] => :script,
-      %w[bold-script] => :scriptbold,
-      %w[fraktur] => :fraktur,
-      %w[sans-serif] => :sans,
-      %w[bold-sans-serif] => :sansbold,
-      %w[sans-serif-italic] => :sansitalic,
-      %w[sans-serif-bold-italic] => :sansbolditalic,
-      %w[monospace] => :monospace,
-    }.freeze
-
     def postprocess(result, filename, _dir)
       result = from_xhtml(cleanup(to_xhtml(textcleanup(result))))
       toHTML(result, filename)
@@ -67,34 +40,8 @@ module IsoDoc::HtmlFunction
       )
     end
 
-    # hotfix for #238, convert mathvariant text into associated plain chars
     def mathml(docxml)
-      docxml.xpath("//m:math", MATHML).each do |elem|
-        mathml1(elem)
-      end
-    end
-
-    def mathml1(base_elem)
-      MATHVARIANT_SPECIAL_CASE_MAPPINGS_1
-        .merge(MATHVARIANT_SPECIAL_CASE_MAPPINGS_2)
-        .merge(MATHVARIANT_TO_PLANE_MAPPINGS)
-        .each_pair do |mathvariant_list, plain_font|
-          xpath = mathvariant_list
-            .map { |variant| "//*[@mathvariant = '#{variant}']" }
-            .join
-          base_elem.xpath(xpath).each do |elem|
-            toPlane(elem, plain_font)
-          end
-        end
-    end
-
-    def toPlane(elem, font)
-      elem.traverse do |n|
-        next unless n.text?
-
-        n.replace(Plane1Converter.conv(HTMLEntities.new.decode(n.text), font))
-      end
-      elem
+      IsoDoc::HtmlFunction::MathvariantToPlain.new(docxml).convert
     end
 
     def htmlstylesheet
@@ -129,7 +76,7 @@ module IsoDoc::HtmlFunction
       auth = docxml.at("//div[@id = 'boilerplate-#{klass}' or @class = 'boilerplate-#{klass}']")
       auth&.xpath(".//h1[not(text())] | .//h2[not(text())]")&.each { |h| h.remove }
       auth&.xpath(".//h1 | .//h2")&.each { |h| h["class"] = "IntroTitle" }
-      dest && auth && dest.replace(auth.remove)
+      dest and auth and dest.replace(auth.remove)
     end
 
     def authority_cleanup(docxml)
@@ -170,7 +117,7 @@ module IsoDoc::HtmlFunction
 
     # needs to be same output as toclevel
     def html_toc(docxml)
-      (idx = docxml.at("//div[@id = 'toc']")) || (return docxml)
+      (idx = docxml.at("//div[@id = 'toc']")) or (return docxml)
       toc = "<ul>"
       path = toclevel_classes.map do |l|
         "//main//#{l}[not(@class = 'TermNum')][not(@class = 'noTOC')][text()]"
@@ -206,9 +153,8 @@ module IsoDoc::HtmlFunction
     def image_suffix(i)
       type = i["mimetype"]&.sub(%r{^[^/*]+/}, "")
       matched = /\.(?<suffix>[^. \r\n\t]+)$/.match i["src"]
-      return type if type && !type.empty?
-      return matched[:suffix] if !matched.nil? && matched[:suffix]
-
+      type and !type.empty? and return type
+      !matched.nil? and matched[:suffix] and return matched[:suffix]
       "png"
     end
 
@@ -265,7 +211,7 @@ module IsoDoc::HtmlFunction
     def footnote_backlinks(docxml)
       seen = {}
       docxml.xpath('//a[@class = "FootnoteRef"]').each_with_index do |x, i|
-        seen[x["href"]] && next || (seen[x["href"]] = true)
+        seen[x["href"]] and next or seen[x["href"]] = true
         fn = docxml.at(%<//*[@id = '#{x['href'].sub(/^#/, '')}']>) || next
         footnote_backlinks1(x, fn)
         x["id"] ||= "fnref:#{i + 1}"
