+++ spec/isodoc/postproc_spec.rb
@@ -617,45 +617,168 @@ TOCLEVEL
 
   end
 
-  it "moves images in HTML" do
-    FileUtils.rm_f "test.html"
-    FileUtils.rm_rf "test_htmlimages"
-    input = <<~INPUT
-      <?xml version="1.0" encoding="UTF-8"?>
-      <iso-standard xmlns="https://www.metanorma.org/ns/iso" type="semantic" version="1.5.14">
-        <sections>
-          <clause id="_clause" inline-header="false" obligation="normative">
-            <title>Clause</title>
-            <p id="_20514f5a-9f86-454e-b6ce-927f65ba6441">
-              <stem type="MathML"><math xmlns="http://www.w3.org/1998/Math/MathML">
+  describe "mathvariant to plain" do
+    context "when `mathvariant` attr equal to `script`" do
+      it "converts mathvariant text chars into associated plain chars" do
+        FileUtils.rm_f "test.html"
+        FileUtils.rm_rf "test_htmlimages"
+        input = <<~INPUT
+          <?xml version="1.0" encoding="UTF-8"?>
+          <iso-standard xmlns="https://www.metanorma.org/ns/iso" type="semantic" version="1.5.14">
+            <sections>
+              <clause id="_clause" inline-header="false" obligation="normative">
+                <title>Clause</title>
+                <p id="_20514f5a-9f86-454e-b6ce-927f65ba6441">
+                  <stem type="MathML"><math xmlns="http://www.w3.org/1998/Math/MathML">
+                      <mi>x</mi>
+                      <mo>=</mo>
+                      <mstyle mathvariant="script">
+                        <mi>l</mi>
+                      </mstyle>
+                      <mo>+</mo>
+                      <mn>1</mn>
+                    </math></stem>
+                </p>
+              </clause>
+            </sections>
+          </iso-standard>
+        INPUT
+        output = <<~OUTPUT
+          <main class="main-section"><button onclick="topFunction()" id="myBtn" title="Go to top">Top</button>
+            <p class="zzSTDTitle1"></p>
+            <div id="_clause">
+              <h1>Clause</h1>
+              <p id="_20514f5a-9f86-454e-b6ce-927f65ba6441">
+              <span class="stem"><math xmlns="http://www.w3.org/1998/Math/MathML">
                   <mi>x</mi>
                   <mo>=</mo>
                   <mstyle mathvariant="script">
-                    <mi>l</mi>
+                    <mi>&#x1D4C1;</mi>
                   </mstyle>
                   <mo>+</mo>
                   <mn>1</mn>
-                </math></stem>
+                </math></span>
             </p>
-          </clause>
-        </sections>
-      </iso-standard>
-    INPUT
-    output = <<~OUTPUT
-      <!DOCTYPE html>
-      <math xmlns="http://www.w3.org/1998/Math/MathML">
-        <mi>x</mi>
-        <mo>=</mo>
-        <mstyle mathvariant="script">
-          <mi>&#x1D4C1;</mi>
-        </mstyle>
-        <mo>+</mo>
-        <mn>1</mn>
-      </math>
-    OUTPUT
-    IsoDoc::HtmlConvert.new({}).convert("test", input, false)
-    html = File.read("test.html")
-    expect(html).to(be_equivalent_to(output))
+            </div>
+          </main>
+        OUTPUT
+        IsoDoc::HtmlConvert.new({}).convert("test", input, false)
+        html = File.read("test.html")
+                .sub(/^.*<main class="main-section">/m, '<main class="main-section">')
+                .sub(%r{</main>.*$}m, "</main>")
+        expect(html).to(be_equivalent_to(output))
+      end
+    end
+
+    context "when complex `mathvariant` combinations" do
+      it "converts mathvariant text chars into associated plain chars" do
+        FileUtils.rm_f "test.html"
+        FileUtils.rm_rf "test_htmlimages"
+        input = <<~INPUT
+          <?xml version="1.0" encoding="UTF-8"?>
+          <iso-standard xmlns="https://www.metanorma.org/ns/iso" type="semantic" version="1.5.14">
+            <sections>
+              <clause id="_clause" inline-header="false" obligation="normative">
+                <title>Clause</title>
+                <p id="_20514f5a-9f86-454e-b6ce-927f65ba6441">
+                  <stem type="MathML">
+                    <math xmlns="http://www.w3.org/1998/Math/MathML">
+                      <mstyle mathvariant="sans-serif">
+                        <mfrac>
+                          <mrow>
+                            <mrow>
+                              <mi>n</mi>
+                              <mfenced open="(" close=")">
+                                <mrow>
+                                  <mstyle mathvariant="bold">
+                                    <mrow>
+                                      <mi>n</mi>
+                                      <mo>+</mo>
+                                      <mn>1</mn>
+                                      <mo>+</mo>
+                                      <mstyle mathvariant="italic">
+                                        <mi>x</mi>
+                                      </mstyle>
+                                    </mrow>
+                                  </mstyle>
+                                </mrow>
+                              </mfenced>
+                            </mrow>
+                          </mrow>
+                          <mrow>
+                            <mstyle mathvariant="bold">
+                              <mrow>
+                                <mi>y</mi>
+                                <mo>+</mo>
+                                <mstyle mathvariant="fraktur">
+                                  <mi>z</mi>
+                                </mstyle>
+                              </mrow>
+                            </mstyle>
+                          </mrow>
+                        </mfrac>
+                      </mstyle>
+                    </math>
+                  </stem>
+                </p>
+              </clause>
+            </sections>
+          </iso-standard>
+        INPUT
+        output = <<~OUTPUT
+          <main class="main-section"><button onclick="topFunction()" id="myBtn" title="Go to top">Top</button>
+            <p class="zzSTDTitle1"></p>
+            <div id="_clause">
+              <h1>Clause</h1>
+              <p id="_20514f5a-9f86-454e-b6ce-927f65ba6441">
+              <span class="stem"><math xmlns="http://www.w3.org/1998/Math/MathML">
+                  <mstyle mathvariant="sans-serif">
+                    <mfrac>
+                      <mrow>
+                        <mrow>
+                          <mi>&#x1D5C7;</mi>
+                          <mfenced open="(" close=")">
+                            <mrow>
+                              <mstyle mathvariant="bold">
+                                <mrow>
+                                  <mi>&#x1D5FB;</mi>
+                                  <mo>+</mo>
+                                  <mn>&#x1D7ED;</mn>
+                                  <mo>+</mo>
+                                  <mstyle mathvariant="italic">
+                                    <mi>&#x1D66D;</mi>
+                                  </mstyle>
+                                </mrow>
+                              </mstyle>
+                            </mrow>
+                          </mfenced>
+                        </mrow>
+                      </mrow>
+                      <mrow>
+                        <mstyle mathvariant="bold">
+                          <mrow>
+                            <mi>&#x1D606;</mi>
+                            <mo>+</mo>
+                            <mstyle mathvariant="fraktur">
+                              <mi>&#x1D59F;</mi>
+                            </mstyle>
+                          </mrow>
+                        </mstyle>
+                      </mrow>
+                    </mfrac>
+                  </mstyle>
+                </math></span>
+            </p>
+            </div>
+          </main>
+        OUTPUT
+        IsoDoc::HtmlConvert.new({}).convert("test", input, false)
+        html = File.read("test.html")
+                .sub(/^.*<main class="main-section">/m, '<main class="main-section">')
+                .sub(%r{</main>.*$}m, "</main>")
+        expect(html).to(be_equivalent_to(output))
+      end
+    end
   end
 
   it "moves images in HTML with no file suffix" do
